Added check for authorization
Added check for authorization in the methods that needed it. Added the check to the service methods and a catch that handles this error in the controllers. That way I prevent the user seeing another user's information.

+++ src/main/java/com/example/petpal/business/impl/PetServiceImpl.java
@@ -3,6 +3,7 @@ package com.example.petpal.business.impl;
 import com.example.petpal.business.IPetService;
 import com.example.petpal.business.domain.*;
 import com.example.petpal.business.exception.*;
+import com.example.petpal.configuration.security.token.IAccessToken;
 import com.example.petpal.persistence.IBreedRepository;
 import com.example.petpal.persistence.IPetRepository;
 import com.example.petpal.persistence.IUserRepository;
@@ -15,6 +16,7 @@ import java.time.LocalDate;
 import java.time.ZoneId;
 import java.util.Date;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -24,19 +26,27 @@ public class PetServiceImpl implements IPetService {
     private final IBreedRepository breedRepository;
     private final IVaccinationRepository vaccinationRepository;
     private final IUserRepository userRepository;
+    private final IAccessToken requestAccessToken;
 
     @Override
-    public Optional<Pet> getPet(Long petId) {
+    public Optional<Pet> getPet(Long petId) throws UnauthorizedDataAccessException {
+        Optional<User> owner = userRepository.getUserByPetId(petId);
+        if(owner.isEmpty() || !Objects.equals(requestAccessToken.getUserId(), owner.get().getId())) {
+            throw new UnauthorizedDataAccessException();
+        }
+
         return petRepository.getPet(petId);
     }
 
     @Transactional
     @Override
-    public Long createPet(Pet pet, Long breedId, List<Long> vaccinationRecordsIds, Long userId) throws InvalidVaccinationException, InvalidBreedException, InvalidUserException, CreationFailException {
+    public Long createPet(Pet pet, Long breedId, List<Long> vaccinationRecordsIds, Long userId) throws InvalidVaccinationException, InvalidBreedException, InvalidUserException, CreationFailException, UnauthorizedDataAccessException {
         try {
             User user = userRepository.getUserById(userId)
                     .orElseThrow(() -> new InvalidUserException(userId));
-
+            if(!Objects.equals(requestAccessToken.getUserId(), user.getId())) {
+                throw new UnauthorizedDataAccessException();
+            }
             Breed breed = breedRepository.getBreedById(breedId)
                     .orElseThrow(() -> new InvalidBreedException(breedId));
 
@@ -59,7 +69,7 @@ public class PetServiceImpl implements IPetService {
             }
 
             return petId;
-        } catch (InvalidUserException | InvalidBreedException | InvalidVaccinationException e) {
+        } catch (InvalidUserException | InvalidBreedException | InvalidVaccinationException | UnauthorizedDataAccessException e) {
             throw e;
         } catch (Exception e) {
             throw new CreationFailException("Failed to create pet and associated vaccination records." + e.getMessage());
@@ -67,10 +77,14 @@ public class PetServiceImpl implements IPetService {
     }
 
     @Override
-    public void updatePet(Pet pet, Long breedId) throws InvalidPetException, InvalidBreedException {
+    public void updatePet(Pet pet, Long breedId) throws InvalidPetException, InvalidBreedException, UnauthorizedDataAccessException {
         Pet existingPet = petRepository.getPet(pet.getId())
                 .orElseThrow(() -> new InvalidPetException(pet.getId()));
 
+        Optional<User> owner = userRepository.getUserByPetId(pet.getId());
+        if(owner.isEmpty() || !Objects.equals(requestAccessToken.getUserId(), owner.get().getId())) {
+            throw new UnauthorizedDataAccessException();
+        }
         Breed breed = breedRepository.getBreedById(breedId)
                 .orElseThrow(() -> new InvalidBreedException(breedId));
 
@@ -85,7 +99,12 @@ public class PetServiceImpl implements IPetService {
     }
 
     @Override
-    public boolean deletePet(Long petId) {
+    public boolean deletePet(Long petId) throws UnauthorizedDataAccessException {
+        Optional<User> owner = userRepository.getUserByPetId(petId);
+        if(owner.isEmpty() || !Objects.equals(requestAccessToken.getUserId(), owner.get().getId())) {
+            throw new UnauthorizedDataAccessException();
+        }
+
         return petRepository.deletePet(petId);
     }
